android: Allow configuration of client identity for all authentication types
This replaces the drop-down box to select certificate identities with a
text field (in the advanced setting) with auto-completion for SANs
contained in the certificate.

The field is always shown and allows using an identity different from
the username for EAP authentication (e.g. to configure a more complete
identity to select a specific config on the server).

 #3134