vici: List additional information for deleted CHILD_SAs
If a CHILD_SA is terminated, the updown event is triggered after the
CHILD_SA is set to state CHILD_DELETED, so no usage stats or detail
information like SPIs were reported.  However, when an IKEv2 SA is
terminated the updown event for its children is triggered without
changing the state first, i.e. they usually remain in state INSTALLED
and detailed data was reported in the event (IKEv1 CHILD_SAs are
terminated individually, i.e. with state change and no extra data so
far).

 #3198